backend/requirements.txt
flask
google-genai>=0.1.0
python-dotenv
requests
duckduckgo-search
redis